requirements.txt
numpy==1.22
pandas==1.1
torch==1.12
scipy==1.9.1
transformers==4.11.3
pytorch-lightning==1.6.4
huggingface_hub
seaborn
mpl-scatter-density
astropy
gitpython
biotite==0.34
requests
bert-embedding
torch
torch.utils.data
sklearn.model_selection
tabulate
tqdm
random
ksvd
Sparselandtools